+++ components/ui/dropdown-menu.tsx
@@ -18,8 +18,7 @@ const DropdownMenuContent = React.forwardRef<
       ref={ref}
       sideOffset={sideOffset}
       className={cn(
-        "z-50 min-w-[18rem] overflow-hidden rounded-xl border border-neutral-200 bg-white p-1 shadow-elevated",
-        "data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95",
+        "z-[100] min-w-[18rem] overflow-hidden rounded-xl border border-neutral-200 bg-white p-1 shadow-elevated",
         className
       )}
       {...props}